@@ -27,25 +27,27 @@
         </developer>         
     </developers>
 	<properties>
-		<java.version>11</java.version>
+		<java.version>17</java.version>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<subethasmtp.version>3.1.7</subethasmtp.version>
 		<commons-net.version>3.9.0</commons-net.version>
 		<snakeyaml.version>1.33</snakeyaml.version>
+		<commons-io.version>2.11.0</commons-io.version>
+		<javax-mail.version>1.6.2</javax-mail.version>
 	</properties>   
 	<dependencyManagement>
 		<dependencies>											
 			<dependency>
 				<groupId>org.springframework.boot</groupId>
 				<artifactId>spring-boot-dependencies</artifactId>
-				<version>2.7.10</version>
+				<version>3.1.0</version>
 				<type>pom</type>
 				<scope>import</scope>
 			</dependency>				
 			<dependency>
 				<groupId>org.springframework.cloud</groupId>
 				<artifactId>spring-cloud-starter-parent</artifactId>
-				 <version>2021.0.6</version>
+				 <version>2022.0.3</version>
 				<type>pom</type>
 				<scope>import</scope>		
 			</dependency>											
