@@ -5,7 +5,7 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.7.12</version>
+	    <version>3.1.0</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>com.example</groupId>
@@ -14,9 +14,10 @@
 	<name>smtp-sink</name>
 	<description>Application to read messages for SMTP MQ channel.</description>
 	<properties>
-		<java.version>11</java.version>
-		<spring-cloud.version>2021.0.6</spring-cloud.version>
+		<java.version>17</java.version>
+		<spring-cloud.version>2022.0.3</spring-cloud.version>
 		<snakeyaml.version>1.33</snakeyaml.version>
+		<commons-io.version>2.11.0</commons-io.version>	
 	</properties>
 	<dependencies>
 		<dependency>
@@ -25,37 +26,21 @@
 		</dependency>
 	    <dependency>
 	        <groupId>org.springframework.cloud</groupId>
-	        <artifactId>spring-cloud-starter-stream-rabbit</artifactId>
-            <exclusions>
-               <exclusion>
-                  <groupId>org.springframework</groupId>
-                  <artifactId>spring-web</artifactId>
-               </exclusion>
-            </exclusions>		        
-	    </dependency>		
+	        <artifactId>spring-cloud-starter-stream-rabbit</artifactId>	        
+	    </dependency>
+	    <dependency>
+			<groupId>org.springframework.cloud</groupId>
+			<artifactId>spring-cloud-stream-binder-kafka</artifactId>
+	    </dependency>		    
 		<dependency>
-			<groupId>org.nhind</groupId>
-			<artifactId>direct-common</artifactId>
-			<version>8.0.0</version>
-            <exclusions>
-               <exclusion>
-                  <groupId>org.springframework.cloud</groupId> 
-                  <artifactId>spring-cloud-starter-openfeign</artifactId>
-               </exclusion>
-               <exclusion>
-                  <groupId>org.apache.mina</groupId> 
-                  <artifactId>mina-core</artifactId>
-               </exclusion>
-               <exclusion>
-                  <groupId>org.jdom</groupId> 
-                  <artifactId>jdom</artifactId>
-               </exclusion>
-               <exclusion>
-                  <groupId>org.apache.commons</groupId>
-                  <artifactId>commons-text</artifactId>
-               </exclusion>
-            </exclusions>				
-		</dependency>	
+		    <groupId>org.springframework.boot</groupId>
+		    <artifactId>spring-boot-starter-mail</artifactId>
+		</dependency>		    
+	    <dependency>
+	    	<groupId>commons-io</groupId>
+	    	<artifactId>commons-io</artifactId>
+	    	<version>${commons-io.version}</version>
+	    </dependency>	    	    			
 		<dependency>
 			<groupId>org.projectlombok</groupId>
 			<artifactId>lombok</artifactId>
@@ -66,13 +51,11 @@
 			<artifactId>spring-boot-starter-test</artifactId>
 			<scope>test</scope>
 		</dependency>		
-		<dependency>
-			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-stream</artifactId>
-			<scope>test</scope>
-			<classifier>test-binder</classifier>
-			<type>test-jar</type>
-		</dependency>
+	    <dependency>
+	       <groupId>org.springframework.cloud</groupId>
+	       <artifactId>spring-cloud-stream-test-binder</artifactId>
+	       <scope>test</scope>
+	    </dependency>	
 	</dependencies>
 	<dependencyManagement>
 		<dependencies>
