@@ -2,21 +2,22 @@ package com.java.example.smtpmq.gateway.task;
 
 import java.util.Arrays;
 import java.util.Calendar;
+import java.util.Properties;
 
-import javax.mail.Message;
-import javax.mail.Session;
-import javax.mail.internet.InternetAddress;
-import javax.mail.internet.MimeMessage;
+import jakarta.mail.Message;
+import jakarta.mail.Session;
+import jakarta.mail.internet.InternetAddress;
+import jakarta.mail.internet.MimeMessage;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.nhindirect.common.mail.SMTPMailMessage;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
+import com.java.example.smtpmq.gateway.server.SMTPMailMessage;
 import com.java.example.smtpmq.gateway.streams.SmtpGatewayMessageSource;
 
 @ConditionalOnProperty(name="smtpmqgateway.loadgen.rate", matchIfMissing=false)
@@ -37,7 +38,7 @@ public class SimulatedLoadTask
 	@Scheduled(fixedRateString = "${smtpmqgateway.loadgen.rate}", initialDelayString = "${smtpmqgateway.loadgen.initialdelay:5000}")
 	public void sendMessages() throws Exception 
 	{
-		final MimeMessage msg = new MimeMessage((Session)null);
+		final MimeMessage msg = new MimeMessage(Session.getDefaultInstance(new Properties()));
 		msg.setFrom(new InternetAddress(sender));
 		msg.setRecipient(Message.RecipientType.TO, new InternetAddress(recipient));
 		msg.setSentDate(Calendar.getInstance().getTime());
